Nav bar change

+++ activismhub.py
@@ -858,14 +858,14 @@ def verifyEmailText(email,hash):
     link=f"{SERVER_NAME}/verifyEmail?e={email}&h={hash}"
     html=f"""\
         <html>
-          <body>
-            <p>Hello,<br><br>
-               Welcome to ACTivism Hub!<br>
-               Please verify your email by clicking <a href={link}>here</a>.<br><br>
-               Best,<br>
-               The ACTivism Hub Team
-            </p>
-          </body>
+            <body>
+                <p>Hello,<br><br>
+                    Welcome to ACTivism Hub!<br>
+                    Please verify your email by clicking <a href={link}>here</a>.<br><br>
+                    Best,<br>
+                    The ACTivism Hub Team
+                </p>
+            </body>
         </html>
         """
     text = f"""\
@@ -930,18 +930,22 @@ def preparePasswordReset():
    club_email=request.form['club_email']
    #Get clubID for that email
    cursor = mysql.connection.cursor()
-   cursor.execute('''SELECT clubID, activation_hash FROM %s WHERE club_email=%%s'''%(CLUB_TABLE,),(club_email,))
+   cursor.execute('''SELECT clubID, activation_hash, club_approved FROM %s WHERE club_email=%%s'''%(CLUB_TABLE,),(club_email,))
    result = cursor.fetchall()
    #if that email isn't in the database
    if len(result) == 0:
         return forgotPassword(club_email+" is not associated with an account.")
-   #Prepare text and send email
+   #check if club approved
    clubID=result[0]['clubID']
-   activation_hash = result[0]['activation_hash']
-   texts = resetPasswordText(club_email,activation_hash)
-   sendEmail(club_email,texts['html'],texts['text'],"Reset your password")
-   #TODO - where to return when email sent? Add pop up
-   return index("An email has been sent to "+club_email+" with a reset link.")
+   if result[0]['club_approved']:
+       #Prepare text and send email
+       activation_hash = result[0]['activation_hash']
+       texts = resetPasswordText(club_email,activation_hash)
+       sendEmail(club_email,texts['html'],texts['text'],"Reset your password")
+       #TODO - where to return when email sent? Add pop up
+       return index("An email has been sent to "+club_email+" with a reset link.")
+   else:
+      return forgotPassword(club_email+" has not yet been approved to make an account.")
 
 
 #route for link clicked in email
@@ -977,15 +981,26 @@ def doPasswordReset():
     #get info from form
     password = request.form['password']
     clubID = request.form['clubID']
-    #hash password
-    saltedPassword = password + salt
-    password = hashlib.sha256(saltedPassword.encode()).hexdigest()
-    #update password in database, set club to active if not already (this counts as email verification)
-    cursor = mysql.connection.cursor()
-    cursor.execute('''UPDATE %s SET password=%%s,email_activated=1 WHERE clubID=%%s'''%(CLUB_TABLE,),(password,clubID))
-    mysql.connection.commit()
-    #render login page
-    return login_page("Password successfully reset.")
+    #check if club approved
+    cursor.execute('''SELECT club_approved,club_email FROM %s WHERE clubID=%%s'''%(CLUB_TABLE,),(clubID,))
+    result = cursor.fetchall()
+    #if no club with that id
+    if len(result)==0:
+        return forgotPassword(result[0]['club_email']+" is not associated with an account.")
+    #if club approved
+    club_approved=result[0]['club_approved']
+    if club_approved:
+        #hash password
+        saltedPassword = password + salt
+        password = hashlib.sha256(saltedPassword.encode()).hexdigest()
+        #update password in database, set club to active if not already (this counts as email verification)
+        cursor = mysql.connection.cursor()
+        cursor.execute('''UPDATE %s SET password=%%s,email_activated=1 WHERE clubID=%%s'''%(CLUB_TABLE,),(password,clubID))
+        mysql.connection.commit()
+        #render login page
+        return login_page("Password successfully reset.")
+    else:
+        return login_page(result[0]['club_email']+" has not yet been approved to make an account.")
 
 
 #returns a dict with the html and plain text versions of a reset password email
@@ -1080,8 +1095,11 @@ def requestApprovalTexts(club_info):
 def approveClub():
     #get clubID
     clubID=request.args.get('id')
-    #get info for that club
+    #set as approved
     cursor = mysql.connection.cursor()
+    cursor.execute('''UPDATE %s SET club_approved=1 WHERE clubID=%%s'''%(CLUB_TABLE,),(clubID,))
+    mysql.connection.commit()
+    #get info for that club
     cursor.execute('''SELECT * FROM %s WHERE clubID=%%s'''%(CLUB_TABLE,),(clubID,))
     club_info = cursor.fetchall()[0]
     #send verification email
@@ -1093,27 +1111,31 @@ def approveClub():
 
 @app.route("/denyClub")
 def denyClub():
-    #NOTE - add check for if club already approved?
     #get clubID
     clubID=request.args.get('id')
-    #get admin email
+    #check if club was already approved
     cursor = mysql.connection.cursor()
-    cursor.execute('''SELECT web_admin_email FROM %s'''%(ADMIN_TABLE))
-    #NOTE - this assumes there is one, and only gets the first one - is this what we want??
-    admin_email=cursor.fetchall()[0]['web_admin_email']
     #get info for club
     cursor.execute('''SELECT * FROM %s WHERE clubID=%%s'''%(CLUB_TABLE,),(clubID,))
     #NOTE- add check? Would only trigger if they approved club, the club deleted itself, then they clicked deny
     club_info = cursor.fetchall()[0]
-    #remove club from database
-    cursor.execute('''DELETE FROM %s WHERE clubID=%%s'''%(CLUB_TABLE,),(clubID))
-    cursor.connection.commit()
-    #send email to club informing them of denial
-    texts=clubDeniedTexts(admin_email)
-    sendEmail(club_info['club_email'],texts['html'],texts['text'],"Request to make club account denied")
-    #load home page with message that club was denied
-    return index(club_info['club_name']+'''\'s request to make an account was denied. An email has been sent to the club
-        to inform them of their denial.''')
+    club_approved=club_info['club_approved']
+    if not club_approved:
+        #get admin email
+        cursor.execute('''SELECT web_admin_email FROM %s'''%(ADMIN_TABLE))
+        #NOTE - this assumes there is one, and only gets the first one - is this what we want??
+        admin_email=cursor.fetchall()[0]['web_admin_email']
+        #remove club from database
+        cursor.execute('''DELETE FROM %s WHERE clubID=%%s'''%(CLUB_TABLE,),(clubID))
+        cursor.connection.commit()
+        #send email to club informing them of denial
+        texts=clubDeniedTexts(admin_email)
+        sendEmail(club_info['club_email'],texts['html'],texts['text'],"Request to make club account denied")
+        #load home page with message that club was denied
+        return index(club_info['club_name']+'''\'s request to make an account was denied. An email has been sent to the club
+            to inform them of their denial.''')
+    else:
+        return index("Denial failed. "+club_info['club_name']+" was previously approved.")
 
 
 #Prepares text for emails if club was denied an account
